Extracted components to separate files
