+++ src/views/TreasuryDashboard/components/Graph/Graph.js
@@ -1,0 +1,198 @@
+import Chart from "src/components/Chart/Chart.jsx";
+import { useTheme } from "@material-ui/core/styles";
+import { trim, formatCurrency } from "../../../../helpers";
+import { useTreasuryMetrics } from "../../hooks/useTreasuryMetrics";
+import { useTreasuryRebases } from "../../hooks/useTreasuryRebases";
+import { bulletpoints, tooltipItems, tooltipInfoMessages, itemType } from "../../treasuryData.js";
+
+export const Graph = ({ children }) => <>{children}</>;
+
+export const TotalValueDepositedGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  return (
+    <Chart
+      type="area"
+      data={data}
+      itemType={itemType.dollar}
+      itemNames={tooltipItems.tvl}
+      dataKey={["totalValueLocked"]}
+      headerText="Total Value Deposited"
+      stopColor={[["#768299", "#98B3E9"]]}
+      bulletpointColors={bulletpoints.tvl}
+      infoTooltipMessage={tooltipInfoMessages.tvl}
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      headerSubText={`${data && formatCurrency(data[0].totalValueLocked)}`}
+    />
+  );
+};
+
+export const MarketValueGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  return (
+    <Chart
+      stopColor={[
+        ["#F5AC37", "#EA9276"],
+        ["#768299", "#98B3E9"],
+        ["#DC30EB", "#EA98F1"],
+        ["#8BFF4D", "#4C8C2A"],
+      ]}
+      dataKey={[
+        "treasuryDaiMarketValue",
+        "treasuryFraxMarketValue",
+        "treasuryWETHMarketValue",
+        "treasuryXsushiMarketValue",
+      ]}
+      data={data}
+      type="stack"
+      itemType={itemType.dollar}
+      itemNames={tooltipItems.coin}
+      bulletpointColors={bulletpoints.coin}
+      headerText="Market Value of Treasury Assets"
+      infoTooltipMessage={tooltipInfoMessages.mvt}
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      headerSubText={`${data && formatCurrency(data[0].treasuryMarketValue)}`}
+    />
+  );
+};
+
+export const RiskFreeValueGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  return (
+    <Chart
+      stopColor={[
+        ["#F5AC37", "#EA9276"],
+        ["#768299", "#98B3E9"],
+        ["#000", "#fff"],
+        ["#000", "#fff"],
+      ]}
+      data={data}
+      type="stack"
+      format="currency"
+      itemType={itemType.dollar}
+      itemNames={tooltipItems.coin}
+      bulletpointColors={bulletpoints.coin}
+      infoTooltipMessage={tooltipInfoMessages.rfv}
+      headerText="Risk Free Value of Treasury Assets"
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      dataKey={["treasuryDaiRiskFreeValue", "treasuryFraxRiskFreeValue"]}
+      headerSubText={`${data && formatCurrency(data[0].treasuryRiskFreeValue)}`}
+    />
+  );
+};
+
+export const ProtocolOwnedLiquidityGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  return (
+    <Chart
+      isPOL
+      type="area"
+      data={data}
+      dataFormat="percent"
+      itemNames={tooltipItems.pol}
+      itemType={itemType.percentage}
+      dataKey={["treasuryOhmDaiPOL"]}
+      bulletpointColors={bulletpoints.pol}
+      infoTooltipMessage={tooltipInfoMessages.pol}
+      headerText="Protocol Owned Liquidity OHM-DAI"
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      headerSubText={`${data && trim(data[0].treasuryOhmDaiPOL, 2)}% `}
+      stopColor={[["rgba(128, 204, 131, 1)", "rgba(128, 204, 131, 0)"]]}
+    />
+  );
+};
+
+export const OHMStakedGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  const staked =
+    data &&
+    data
+      .map(metric => ({
+        staked: (metric.sOhmCirculatingSupply / metric.ohmCirculatingSupply) * 100,
+        timestamp: metric.timestamp,
+      }))
+      .filter(metric => metric.staked < 100);
+
+  return (
+    <Chart
+      isStaked
+      type="area"
+      data={staked}
+      dataKey={["staked"]}
+      dataFormat="percent"
+      headerText="OHM Staked"
+      stopColor={[["#55EBC7", "#47ACEB"]]}
+      bulletpointColors={bulletpoints.staked}
+      infoTooltipMessage={tooltipInfoMessages.staked}
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      headerSubText={`${staked && trim(staked[0].staked, 2)}% `}
+    />
+  );
+};
+
+export const APYOverTimeGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryRebases();
+
+  let apy =
+    data &&
+    data
+      .map(entry => ({
+        timestamp: entry.timestamp,
+        apy: Math.pow(parseFloat(entry.percentage) + 1, 365 * 3) * 100,
+      }))
+      .filter(pm => pm.apy < 300000);
+
+  return (
+    <Chart
+      type="line"
+      scale="log"
+      data={apy}
+      dataKey={["apy"]}
+      dataFormat="percent"
+      headerText="APY over time"
+      itemNames={tooltipItems.apy}
+      itemType={itemType.percentage}
+      color={theme.palette.text.primary}
+      bulletpointColors={bulletpoints.apy}
+      stroke={[theme.palette.text.primary]}
+      infoTooltipMessage={tooltipInfoMessages.apy}
+      headerSubText={`${apy && trim(apy[0].apy, 2)}%`}
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+    />
+  );
+};
+
+export const RunwayAvailableGraph = () => {
+  const theme = useTheme();
+  const { data } = useTreasuryMetrics();
+
+  const runway = data && data.filter(metric => metric.runway10k > 5);
+
+  return (
+    <Chart
+      type="line"
+      data={runway}
+      itemType={""}
+      dataFormat="days"
+      dataKey={["runwayCurrent"]}
+      headerText="Runway Available"
+      itemNames={tooltipItems.runway}
+      color={theme.palette.text.primary}
+      stroke={[theme.palette.text.primary]}
+      bulletpointColors={bulletpoints.runway}
+      infoTooltipMessage={tooltipInfoMessages.runway}
+      expandedGraphStrokeColor={theme.palette.graphStrokeColor}
+      headerSubText={`${data && trim(data[0].runwayCurrent, 1)} Days`}
+    />
+  );
+};
