+++ src/views/TreasuryDashboard/hooks/useTreasuryRebases.js
@@ -1,0 +1,14 @@
+import { useQuery } from "react-query";
+import apollo from "src/lib/apolloClient";
+import { rebasesDataQuery } from "../treasuryData.js";
+
+export const useTreasuryRebases = () => {
+  return useQuery("treasury_rebases", async () => {
+    const response = await apollo(rebasesDataQuery);
+
+    console.log("rebases");
+
+    // Transform string values to floats
+    return response.data.rebases;
+  });
+};
